pyproject.toml
[tool.poetry]
name = "router-service"
version = "0.1.0"
description = ""
authors = ["David Hellinga <d.hellinga@student.fontys.nl>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
osmnx = "^1.3.0"
scipy = "^1.10.1"
networkx = "^3.1"
pydantic = "^1.10.7"
flake8-isort = "^6.0.0"
masstransitpython-byq96 = "^0.0.5"
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pyenchant = "^3.2.2"
black = "^23.1.0"
flake8 = "^6.0.0"
yamllint = "^1.30.0"
pylint = "^2.17.1"
flake8-pyproject = "^1.2.3"
flake8-blind-except = "^0.2.1"
flake8-comprehensions = "^3.11.1"
flake8-docstrings = "^1.7.0"
flake8-print = "^5.0.0"
flake8-pytest = "^1.4"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# See https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
profile = "black"

[tool.pytest.ini_options]
markers = [
    "unit",
    "contract",
    "e2e",
]


[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "I003", "I004"]
doctests = true
docstring-convention = "numpy"
import-order-style = "smarkets"
max-complexity = 72

exclude = [
    ".git",
    "__pycache__",
    "bin",
    "include",
    "lib"
]

# D200 - One-line docstring should fit on one line with quotes
# D204 - 1 blank line required after class docstring
# S101 - Use of assert detected
# W503 - line break before binary operator
# E501 - line too long
ignore = ["D200", "D204", "S101", "W503", "E501"]

# W605 - Invalid regex escape character. Ignored because Pandas still uses them.
# per-file-ignores =
